+++ examples-spring-boot/src/main/java/com/github/dbmdz/flusswerk/spring/boot/example/Application.java
@@ -2,8 +2,8 @@ package com.github.dbmdz.flusswerk.spring.boot.example;
 
 import com.github.dbmdz.flusswerk.framework.config.properties.FlusswerkProperties;
 import com.github.dbmdz.flusswerk.framework.engine.Engine;
-import com.github.dbmdz.flusswerk.framework.spring.EnableFlusswerk;
-import com.github.dbmdz.flusswerk.framework.spring.FlusswerkApplication;
+import com.github.dbmdz.flusswerk.framework.EnableFlusswerk;
+import com.github.dbmdz.flusswerk.framework.FlusswerkApplication;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
