--- examples-spring-boot/src/main/java/com/github/dbmdz/flusswerk/spring/boot/example/model/GreetingMixin.java
@@ -1,17 +1,0 @@
-package com.github.dbmdz.flusswerk.spring.boot.example.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(Include.NON_EMPTY)
-public interface GreetingMixin {
-
-  @JsonIgnore
-  int getDeliveryTag();
-
-  @JsonIgnore
-  int getBody();
-}
