+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/FlusswerkApplication.java
@@ -1,4 +1,4 @@
-package com.github.dbmdz.flusswerk.framework.spring;
+package com.github.dbmdz.flusswerk.framework;
 
 import com.github.dbmdz.flusswerk.framework.engine.Engine;
 import org.springframework.boot.CommandLineRunner;
