+++ framework/src/test/java/com/github/dbmdz/flusswerk/framework/messagebroker/RabbitClientTest.java
@@ -15,7 +15,7 @@ import com.github.dbmdz.flusswerk.framework.TestMessage;
 import com.github.dbmdz.flusswerk.framework.exceptions.InvalidMessageException;
 import com.github.dbmdz.flusswerk.framework.model.Envelope;
 import com.github.dbmdz.flusswerk.framework.model.Message;
-import com.github.dbmdz.flusswerk.framework.spring.MessageImplementation;
+import com.github.dbmdz.flusswerk.framework.model.IncomingMessageType;
 import com.rabbitmq.client.AMQP.BasicProperties;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.GetResponse;
@@ -66,7 +66,7 @@ class RabbitClientTest {
   @Test
   void shouldWorkWithCustomMessageType() throws IOException {
     var messageImplementation =
-        new MessageImplementation(TestMessage.class, TestMessageMixin.class);
+        new IncomingMessageType(TestMessage.class, TestMessageMixin.class);
     RabbitClient client = new RabbitClient(messageImplementation, connection);
 
     TestMessage message = new TestMessage("abc123", "should be ignored");
